Add --exclude-aggregation-questions to PW eval
Also add question "difficulty" to preds metadata

PW datasets contain aggregation questions, but add option to exclude them for evaluation.
Evaluation is expensive in time and cost, so it's better to filter them out before evaluating an LLM
than to evaluate an LLM on all questions and then filter out the results.