Use <answer>...</answer> in PhantomEval prompts (#301)
- Change CoT prompts from "The answer is ..." to "<answer>...</answer>"
and accordingly update the CoTAgent parsing.
- Update Nshot parsing to handle optional <think> tags in model outputs.
Some LLMs generate this by default, so skip those.
- `--inf_is_deepseek_r1_model` previously served 2 purposes: switch on
`parse_thinking_answer` kwarg in the PhantomEval pipeline, and add an
additional stop token in vllm. But even without that flag the results
are the same as reported numbers in PhantomWiki arxiv paper. So
completely remove the flag, since the additional token is also not
necessary.

Results on huggingface published `--dataset kilian-group/phantom-wiki-v1
--split_list depth_20_size_50_seed_1`, demonstrating that the PR
implementation and requesting <answer> tags does not materially change
the LLM behavior.

- deepseek produces <think> tags in zeroshot and cot, the numbers are
same as paper. So <think> parsing works.
- llama does not produce <think> tags in zeroshot or cot. With <answer>
tags requested in cot prompting, the performance is slightly worse than
reported in paper

| ('Universe Size', '') | ('Model', '') | ('In-Context',
'\\textsc{Zeroshot}') | ('In-Context', '\\textsc{CoT}') |

|-------------------------|-----------------|----------------------------------------|-----------------------------------|
| 50 | DeepSeek-R1-32B | 42.85 ± 2.88 | 49.58 ± 2.06 |
| 50 | Llama-3.3-70B | 25.95 ± 0.55 | 44.15 ± 2.15 |

---------

Co-authored-by: Albert Gong <ag2435@cornell.edu>